LDW/text09/create_table.py
import psycopg2
import os
from dotenv import load_dotenv

load_dotenv(os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))), '.env'))

DB_HOST = "127.0.0.1"
DB_NAME = "interview_db"
DB_USER = "postgres"
DB_PASS = os.getenv("POSTGRES_PASSWORD", "013579")
DB_PORT = "5432"

try:
    conn = psycopg2.connect(
        host=DB_HOST,
        database=DB_NAME,
        user=DB_USER,
        password=DB_PASS,
        port=DB_PORT
    )
    cur = conn.cursor()
    
    # Create interview_announcement table
    cur.execute('''
        CREATE TABLE IF NOT EXISTS interview_announcement (
            id SERIAL PRIMARY KEY,
            title TEXT NOT NULL,
            deadline TEXT,
            content TEXT,
            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
        )
    ''')
    
    conn.commit()
    print("Table 'interview_announcement' created successfully.")
    
    cur.close()
    conn.close()
except Exception as e:
    print(f"Error: {e}")
